requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.21
django-cors-headers==4.7.0
djangorestframework==3.16.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
kombu==5.5.3
packaging==25.0
promise==2.3
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
